// src/app/models/escola-model.ts
export interface Escola {
    avaliacao: number | null;
    bairro: string;
    cep: string;
    cidade: string;
    email: string;
    estado: string;
    id: number;
    mensalidade: number;
    metodologia: string;
    nome: string;
    numero: string;
    quantidade_alunos: number;
    rua: string;
    telefone: string;
    imagem_url: string

  }